requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
pydantic
pandas
duckduckgo-search
tiktoken